# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cargexo Documentation",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Cargexo API documentation"
        }
    ],
    "paths": {
...
        "/all/customers": {
            "post": {
                "tags": [
                    "Customer"
                ],
                "summary": "get all customers",
                "description": "get all customers",
                "operationId": "get all customers",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "filter": {
                                        "type": "array",
                                        "items": {
                                            "type": "array",
                                            "items": {}
                                        },
                                        "example": [
                                            [
                                                "mc_number",
                                                "=",
                                                2
                                            ],
                                            [
                                                "user_id",
                                                "=",
                                                "2"
                                            ]
                                        ]
                                    },
                                    "columns": {
                                        "type": "array",
                                        "items": {
                                            "type": "array",
                                            "items": {}
                                        },
                                        "example": [
                                            "id",
                                            "website",
                                            "mc_number"
                                        ]
                                    },
                                    "sort": {
                                        "type": "object",
                                        "example": {
                                            "column": "id",
                                            "dir": "desc"
                                        }
                                    },
                                    "skip": {
                                        "type": "string",
                                        "example": "0"
                                    },
                                    "take": {
                                        "type": "string",
                                        "example": "10"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/customers": {
            "post": {
                "tags": [
                    "Customer"
                ],
                "summary": "create customers",
                "description": "create customers",
                "operationId": "create customers",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "personal_phone": {
                                        "type": "integer"
                                    },
                                    "business_phone": {
                                        "type": "integer"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "date": {
                                        "type": "string"
                                    },
                                    "note": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string"
                                    },
                                    "mc_number": {
                                        "type": "string"
                                    },
                                    "dot_number": {
                                        "type": "string"
                                    },
                                    "website": {
                                        "type": "string"
                                    },
                                    "invoive_factoring_approvment": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "invoice_mail": {
                                        "type": "integer",
                                        "example": 0
                                    },
                                    "personal_fax": {
                                        "type": "integer"
                                    },
                                    "business_fax": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/customers/{id}": {
            "post": {
                "tags": [
                    "Customer"
                ],
                "summary": "Get customers By ID",
                "description": "Returns customers full info",
                "operationId": "get customers ByID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "customer ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/update/customers/{id}": {
            "post": {
                "tags": [
                    "Customer"
                ],
                "summary": "update customers",
                "description": "update customers",
                "operationId": "update customers",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Customer ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "personal_phone": {
                                        "type": "integer"
                                    },
                                    "business_phone": {
                                        "type": "integer"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "date": {
                                        "type": "string"
                                    },
                                    "note": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string"
                                    },
                                    "mc_number": {
                                        "type": "string"
                                    },
                                    "dot_number": {
                                        "type": "string"
                                    },
                                    "website": {
                                        "type": "string"
                                    },
                                    "invoive_factoring_approvment": {
                                        "type": "string"
                                    },
                                    "invoice_mail": {
                                        "type": "string"
                                    },
                                    "personal_fax": {
                                        "type": "integer"
                                    },
                                    "business_fax": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/update/customers/{id}/status": {
            "post": {
                "tags": [
                    "Customer"
                ],
                "summary": "update customers status ",
                "description": "update customers status",
                "operationId": "update customers status",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "customersID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/delete/customers/{id}": {
            "delete": {
                "tags": [
                    "Customer"
                ],
                "summary": "delete customers ",
                "description": "delete customers",
                "operationId": "delete customers",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "customer ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "name",
                        "description": "name of the new User",
                        "type": "string",
                        "example": "Ahmed"
                    },
                    "email": {
                        "title": "email",
                        "description": "email of the new User",
                        "type": "string",
                        "example": "email@email.com"
                    },
                    "personal_phone": {
                        "title": "personal_phone",
                        "description": "personal_phone of the new User",
                        "type": "string",
                        "example": "0770123456"
                    },
                    "business_phone": {
                        "title": "business_phone",
                        "description": "business_phone of the new User",
                        "type": "string",
                        "example": "0770123456"
                    },
                    "password": {
                        "title": "password",
                        "description": "password of the new User",
                        "type": "string",
                        "example": "32134"
                    },
                    "type": {
                        "title": "type",
                        "description": "type of the new User",
                        "type": "string",
                        "example": "admin"
                    },
                    "address": {
                        "title": "address",
                        "description": "address of the new User",
                        "type": "string",
                        "example": "Baghdad/alkaradaa 62 Street"
                    },
                    "date": {
                        "title": "date",
                        "description": " signIn date or hiring date ",
                        "type": "string",
                        "example": "2/2/2020"
                    },
                    "status": {
                        "title": "status",
                        "description": "status of the new User",
                        "type": "string",
                        "example": "active"
                    },
                    "is_deleted": {
                        "title": "is_deleted",
                        "description": "is_deleted determain if the user deleted or not",
                        "type": "boolean",
                        "example": "true"
                    },
                    "note": {
                        "title": "note",
                        "description": "note about the user",
                        "type": "boolean",
                        "example": "true"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            }
        },
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}